Rename power_features.h/c to <arch>_power_features.h/c and update GPU builds
