+++ src/variorum/AMD/amd_power_features.c
