+++ src/variorum/AMD/amd_power_features.h
