+++ src/variorum/AMD_GPU/CMakeLists.txt
@@ -5,12 +5,12 @@
 
 set(variorum_amd_gpu_headers
   ${CMAKE_CURRENT_SOURCE_DIR}/instinctGPU.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/power_features.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/amd_gpu_power_features.h
   CACHE INTERNAL "")
 
 set(variorum_amd_gpu_sources
   ${CMAKE_CURRENT_SOURCE_DIR}/instinctGPU.c
-  ${CMAKE_CURRENT_SOURCE_DIR}/power_features.c
+  ${CMAKE_CURRENT_SOURCE_DIR}/amd_gpu_power_features.c
   CACHE INTERNAL "")
 
 include_directories(${CMAKE_CURRENT_SOURCE_DIR} ${variorum_includes})
