+++ src/variorum/AMD_GPU/amd_gpu_power_features.h
