+++ src/variorum/Nvidia_GPU/CMakeLists.txt
@@ -5,12 +5,12 @@
 
 set(variorum_nvidia_headers
   ${CMAKE_CURRENT_SOURCE_DIR}/Volta.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/power_features.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/nvidia_gpu_power_features.h
   CACHE INTERNAL "")
 
 set(variorum_nvidia_sources
   ${CMAKE_CURRENT_SOURCE_DIR}/Volta.c
-  ${CMAKE_CURRENT_SOURCE_DIR}/power_features.c
+  ${CMAKE_CURRENT_SOURCE_DIR}/nvidia_gpu_power_features.c
   CACHE INTERNAL "")
 
 include_directories(${CMAKE_CURRENT_SOURCE_DIR} ${variorum_includes})
